genai-chatbot/requirements.txt
Flask>=3.0,<4
python-dotenv==1.0.1
groq==0.31.0
httpx==0.27.0
ujson==5.10.0